web-animations/timing-model/animation-effects/active-time.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Active time tests</title>
<link rel="help" href="https://w3c.github.io/web-animations/#active-time">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

test(function(t) {
  var div = createDiv(t);
  div.style.lineHeight = '1234px';
  var anim = div.animate({ lineHeight: [ 0, 1 ] }, { delay: 1 });
  anim.currentTime = -1;
  assert_equals(getComputedStyle(div).lineHeight, '1234px');
  anim.currentTime = 1;
  assert_equals(getComputedStyle(div).lineHeight, '1234px');
}, 'Test progress during before and after phase when fill is none');

</script>
</body>
